+++ src/ngx-youtube-player.ts
@@ -1,7 +1,5 @@
 // Public classes.
-export {
-  IPlayerApiScriptOptions, IPlayerOutputs, IPlayerSize
-} from './models';
+export { IPlayerApiScriptOptions, IPlayerOutputs, IPlayerSize } from './models';
 
 export { YoutubePlayerModule } from './modules/ngx-youtube-player.module';
 export { YoutubePlayerComponent } from './modules/youtube-player.component';