+++ CHANGELOG.md
@@ -1,27 +1,38 @@
+## v 6.0.0 (2018/05/30)
+
+* [UPGRADE] - official support for Angular 6, rxjs 6
+* [REFACTOR] - "@types/youtube" is a dependency now
+* [REFACTOR] - Inputs height & width are assigned with the default sizes now
+
 ## v 0.1.0 (2018/01/08)
+
 * [UPGRADE] - official support for Angular 5
 * [UPDATE] - updated repo to ngx-youtube-player
-* [REFACTOR] - added more unit tests and increased coverage 
+* [REFACTOR] - added more unit tests and increased coverage
 
 ## v 0.0.51 (2017/12/29)
-* [FIX] - fixes #27 - youtube player iframe api loaded with each instance 
+
+* [FIX] - fixes #27 - youtube player iframe api loaded with each instance
 
 ## v 0.0.4 (2017/10/03)
+
 * RENAMED npm package to "ngx-youtube-player"
 * updated to be usign Renderer2 for DOM Access to support NodeJS
 * fixes #8
 
 ## v 0.0.3 (2017/05/15)
+
 * updated to work with ng4
 * auto setting for youtube player api protocol (http or https)
 * using angular-library-starter
 * added support for "http" method
 
 ## v 0.0.2 (2017/03/29)
+
 * added examples
 
 ## v 0.0.1 (2017/05/10)
 
 ### Updates
-* First release
 
+* First release